* neighbourhoodLimit is quantized with step 0.2, not 1.0, so it can't be   an int!

+++ plugins/SegmenterPlugin.h
@@ -52,7 +52,7 @@ protected:
     mutable Segmenter* segmenter;
     mutable int hopsize;
     mutable int windowsize;
-    mutable int neighbourhoodLimit; // in sec
+    mutable float neighbourhoodLimit; // in sec
     int nSegmentTypes;
     feature_types featureType;	// 1 = constant-Q, 2 = chroma
